Fix up some windows compilation issues.
These were mostly cases where our previous macros had been casting,
and the values that we were trying to printf were not in fact
uint64_t.

+++ src/lib/time/tvdiff.c
@@ -52,14 +52,14 @@ tv_udiff(const struct timeval *start, const struct timeval *end)
   if (start->tv_usec > TOR_USEC_PER_SEC || start->tv_usec < 0) {
     log_warn(LD_GENERAL, "comparing times on microsecond detail with bad "
              "start tv_usec: %"PRId64 " microseconds",
-             (start->tv_usec));
+             (int64_t)start->tv_usec);
     return LONG_MAX;
   }
 
   if (end->tv_usec > TOR_USEC_PER_SEC || end->tv_usec < 0) {
     log_warn(LD_GENERAL, "comparing times on microsecond detail with bad "
              "end tv_usec: %"PRId64 " microseconds",
-             (end->tv_usec));
+             (int64_t)end->tv_usec);
     return LONG_MAX;
   }
 
@@ -101,14 +101,14 @@ tv_mdiff(const struct timeval *start, const struct timeval *end)
   if (start->tv_usec > TOR_USEC_PER_SEC || start->tv_usec < 0) {
     log_warn(LD_GENERAL, "comparing times on millisecond detail with bad "
              "start tv_usec: %"PRId64 " microseconds",
-             (start->tv_usec));
+             (int64_t)start->tv_usec);
     return LONG_MAX;
   }
 
   if (end->tv_usec > TOR_USEC_PER_SEC || end->tv_usec < 0) {
     log_warn(LD_GENERAL, "comparing times on millisecond detail with bad "
              "end tv_usec: %"PRId64 " microseconds",
-             (end->tv_usec));
+             (int64_t)end->tv_usec);
     return LONG_MAX;
   }
 
@@ -124,7 +124,7 @@ tv_mdiff(const struct timeval *start, const struct timeval *end)
   if (secdiff > (int64_t)(LONG_MAX/1000 - 2) ||
       secdiff < (int64_t)(LONG_MIN/1000 + 1)) {
     log_warn(LD_GENERAL, "comparing times on millisecond detail too far "
-             "apart: %"PRId64 " seconds", (secdiff));
+             "apart: %"PRId64 " seconds", (int64_t)secdiff);
     return LONG_MAX;
   }
